requirements.txt
# Stamp Philatex Processor - Dependencies
# Install with: pip install -r requirements.txt

# =============================================================================
# CORE ML/AI
# =============================================================================
ultralytics>=8.0.0          # YOLOv8/YOLO11 framework
torch>=2.0.0                # PyTorch deep learning
torchvision>=0.15.0         # Vision models & utilities

# AMD GPU Support (Windows)
# Note: For DirectML support, also run:
# pip install torch-directml
torch-directml>=0.2.0       # AMD GPU acceleration on Windows

# =============================================================================
# IMAGE PROCESSING
# =============================================================================
opencv-python>=4.8.0        # Computer vision operations
numpy>=1.24.0               # Numerical computing
pillow>=10.0.0              # Image I/O and manipulation
pillow-heif>=0.13.0         # HEIC/HEIF support for iPhone photos

# =============================================================================
# DUPLICATE DETECTION
# =============================================================================
imagehash>=4.3.0            # Perceptual image hashing

# =============================================================================
# DATABASE
# =============================================================================
# SQLite is built-in to Python, no extra package needed

# =============================================================================
# GUI (PyQt6)
# =============================================================================
PyQt6>=6.5.0                # Qt6 GUI framework
PyQt6-Qt6>=6.5.0            # Qt6 binaries
PyQt6-sip>=13.5.0           # SIP bindings

# =============================================================================
# UTILITIES
# =============================================================================
pyyaml>=6.0                 # YAML config parsing
tqdm>=4.65.0                # Progress bars
pandas>=2.0.0               # Data manipulation
matplotlib>=3.7.0           # Visualization/plotting

# =============================================================================
# OPTIONAL: DEVELOPMENT & TESTING
# =============================================================================
# pytest>=7.0.0             # Testing framework
# black>=23.0.0             # Code formatting
# flake8>=6.0.0             # Linting
# =============================================================================
# BUILD TOOLS
# =============================================================================
pyinstaller>=6.0.0            # EXE generation
